refactor: code split 1. Split code to be more understandable. 2. Reuse the same variables across the entire backend.
