+++ apps/nest-backend/src/configs/project.config.ts
@@ -8,3 +8,11 @@ export const CONFIG_CURRENT_PROJECT_PATH = 'currentProjectPath';
 export const ABSTRACT_REPORT_FILE_NAME = 'abstract.json';
 export const SPECS = 'spec.json';
 export const SETTINGS = 'settings.json';
+export const BROWSER_ARGS = [
+  '--window-size=1440,900',
+  '--no-sandbox',
+  '--disable-setuid-sandbox',
+  '--disable-dev-shm-usage',
+  '--disable-accelerated-2d-canvas',
+  '--disable-gpu',
+];
