+++ apps/nest-backend/src/inspector/inspector-utils.service.ts
@@ -1,0 +1,57 @@
+import { HttpException, Injectable } from '@nestjs/common';
+import {
+  StrictDataLayerEvent,
+  ValidationStrategy,
+} from '../interfaces/dataLayer.interface';
+import {
+  EcommerceEventValidationStrategy,
+  OldGA4EventsValidationStrategy,
+} from './strategy/dataLayer-validation-strategy';
+import { ValidationStrategyType, determineStrategy } from './utilities';
+
+@Injectable()
+export class InspectorUtilsService {
+  private validationStrategies: { [key: string]: ValidationStrategy };
+
+  constructor() {
+    this.validationStrategies = {
+      [ValidationStrategyType.ECOMMERCE]:
+        new EcommerceEventValidationStrategy(),
+      [ValidationStrategyType.OLDGA4EVENTS]:
+        new OldGA4EventsValidationStrategy(),
+    };
+  }
+
+  // return true if the dataLayer is correct
+  // return missing keys if the dataLayer is partially correct
+  // return false if the dataLayer object hasn't been found
+  isDataLayerCorrect(
+    dataLayer: StrictDataLayerEvent[],
+    spec: StrictDataLayerEvent
+  ) {
+    const strategyType = determineStrategy(spec);
+
+    try {
+      switch (strategyType) {
+        case ValidationStrategyType.ECOMMERCE:
+          return this.validationStrategies[strategyType].validateDataLayer(
+            dataLayer,
+            spec
+          );
+        case ValidationStrategyType.OLDGA4EVENTS:
+          return this.validationStrategies[strategyType].validateDataLayer(
+            dataLayer,
+            spec
+          );
+        default:
+          return {
+            passed: false,
+            message: "The test didn't pass",
+            dataLayerSpec: spec,
+          };
+      }
+    } catch (error) {
+      throw new HttpException(`${error.message}`, 500);
+    }
+  }
+}
