+++ apps/nest-backend/src/os/xlsx-report/xlsx-report-single-event.service.ts
@@ -1,18 +1,12 @@
 import { HttpException, Injectable, Logger } from '@nestjs/common';
-import * as ExcelJS from 'exceljs';
+import { FolderPathService } from '../path/folder-path/folder-path.service';
 import { readFileSync } from 'fs';
 import path from 'path';
-import { FolderPathService } from '../path/folder-path/folder-path.service';
-import { FilePathService } from '../path/file-path/file-path.service';
-import { FileService } from '../file/file.service';
+import * as ExcelJS from 'exceljs';
 
 @Injectable()
-export class XlsxReportService {
-  constructor(
-    private folderPathService: FolderPathService,
-    private filePathService: FilePathService,
-    private fileService: FileService
-  ) {}
+export class XlsxReportSingleEventService {
+  constructor(private folderPathService: FolderPathService) {}
   async writeXlsxFile(
     fileName: string,
     sheetName: string,
@@ -105,40 +99,5 @@ export class XlsxReportService {
       Logger.error(error.message, 'XlsxReportService.writeXlsxFile');
       throw new HttpException(error.message, 500);
     }
-  }
-
-  async writeXlsxFileForAllTests(
-    operations: string[],
-    fileName: string,
-    sheetName: string,
-    projectName: string
-  ) {
-    const data = [];
-    for (let i = 0; i < operations.length; i++) {
-      const operation = operations[i];
-      const cachePath = await this.filePathService.getCacheFilePath(
-        projectName,
-        operation
-      );
-      const cache = this.fileService.readJsonFile(cachePath);
-      data.push(cache);
-    }
-
-    const result = data.map((item) => {
-      return {
-        dataLayerResult: item.dataLayerCheckResult,
-        requestCheckResult: item.requestCheckResult,
-        rawRequest: item.rawRequest,
-        destinationUrl: item.destinationUrl,
-      };
-    });
-
-    await this.writeXlsxFile(
-      fileName,
-      sheetName,
-      result,
-      undefined,
-      projectName
-    );
   }
 }
