+++ apps/nest-backend/src/os/xlsx-report/xlsx-report.module.ts
@@ -1,11 +1,18 @@
 import { Module } from '@nestjs/common';
-import { XlsxReportService } from './xlsx-report.service';
+import { XlsxReportGroupEventsService } from './xlsx-report-group-events.service';
 import { PathModule } from '../path/path.module';
 import { FileModule } from '../file/file.module';
 import { FileService } from '../file/file.service';
+import { XlsxReportSingleEventService } from './xlsx-report-single-event.service';
+
+const services = [
+  XlsxReportGroupEventsService,
+  XlsxReportSingleEventService,
+  FileService,
+];
 @Module({
   imports: [PathModule, FileModule],
-  providers: [XlsxReportService, FileService],
-  exports: [XlsxReportService],
+  providers: [...services],
+  exports: [...services],
 })
 export class XlsxReportModule {}
